requirements.txt
fastapi
uvicorn 
sqlalchemy 
pymysql
python-dotenv
tabulate
pyjwt
pytest
moviepy
numpy
requests
pillow==9.5.0
colorlog